Update the notification services image to support the configuration of SMTP transport security using the `PASS_NOTIFICATION_SMTP_TRANSPORT` environment variable.
Valid values are: `SMTP`, `SMTPS`, and `SMTP_TLS`.  Note that `SMTP` will use TLS if the server advertises `STARTTLS` (a. k. a. opportunistic TLS).  When opportunistic TLS is used, Notification Services will *not* validate the SMTP server certificate, making attacks on the transport layer trivial.  `SMTPS` and `SMTP_TLS` will strictly verify the certificate presented by the server.

The default value in this PR is `SMTP`, but for production `SMTP_TLS` is probably preferred.